// frontend/src/enums/TextFieldKeys.ts
export enum PersonalDetailTextFieldKeys {
  FIRST_NAME = "first-name",
  LAST_NAME = "last-name",
  EMAIL = "email",
  PHONE = "phone-number",
  DOB = "date-of-birth",
  ADDRESS = "address",
}

export enum FinanceDetailTextFieldKeys {
  SALARY = "income-salary",
  OTHER = "income-other",
  MORTGAGE = "expenditure-mortgage",
  RENT = "expenditure-rent",
  UTILITIES = "expenditure-utilities",
  TRAVEL = "expenditure-travel",
  FOOD = "expenditure-food",
  LOANS = "debt-payments-loans",
  CREDIT_CARDS = "debt-payments-credit-cards",
}